requirements.txt
openai-whisper
torch
tkinter